feat(vo): schedule, joinCondition, contact 필드 멤버 추가
